refactor(BridgeGame, InputView, OutputView): 패키지를 변경
